+++ web/src/util/index.test.js
@@ -42,7 +42,6 @@ describe('provides default years based on now', () => {
 describe('utility functions', () => {
   const {
     applyToNumbers,
-    arrToObj,
     generateKey,
     getParams,
     stateDateToDisplay,
@@ -115,14 +114,5 @@ describe('utility functions', () => {
     expect(stateDateRangeToDisplay()).toEqual(
       'Date not specified - Date not specified'
     );
-  });
-
-  test('converts an array into an object with array values as object keys', () => {
-    expect(arrToObj(['a', 'b', 'c'])).toEqual({ a: 0, b: 0, c: 0 });
-    expect(arrToObj(['a', 'b', 'c'], 'boop')).toEqual({
-      a: 'boop',
-      b: 'boop',
-      c: 'boop'
-    });
   });
 });
